RDF is being generated, closing #5

+++ shortbol/core/shared/src/main/scala/uk/co/turingatemyhamster/shortbol/ExportShortbol.scala
@@ -3,7 +3,7 @@ package uk.co.turingatemyhamster.shortbol
 import java.io.{File, FileWriter}
 import javax.xml.stream.XMLOutputFactory
 
-import uk.co.turingatemyhamster.shortbol.ops.{Exporter, ShortbolParser}
+import uk.co.turingatemyhamster.shortbol.ops.{Exporter, PrettyPrinter, ShortbolParser}
 import ShortbolParser.POps
 import ast.sugar._
 import com.sun.xml.internal.txw2.output.IndentingXMLStreamWriter
@@ -21,17 +21,33 @@ import scala.io.Source
   */
 object ExportShortbol {
   def main(args: Array[String]): Unit = for(file <- args) {
-    ShortbolParser.SBFile.withPositions(file, Source.fromFile(new File(file)).mkString) match {
-      case s : Success[ast.SBFile] =>
-        val (c, v) = s.value.eval.run(Fixture.configuredContext)
-        val doc = Exporter[datatree.ast.AstDatatree](c).apply(v)
-        val rdfIo = implicitly[RdfIo[datatree.ast.AstDatatree]]
-        val xmlWriter = new IndentingXMLStreamWriter(
-          XMLOutputFactory.newInstance.createXMLStreamWriter(
-            new FileWriter(s"$file.rdf")))
-        rdfIo.write(xmlWriter, doc)
-      case e : Failure =>
-        System.err.println(s"Problem parsing $file")
+    println(s"Processing file $file")
+    val f = new File(file)
+    if(f.exists()) {
+      ShortbolParser.SBFile.withPositions(file, Source.fromFile(new File(file)).mkString) match {
+        case s : Success[ast.SBFile] =>
+          println("Loaded:")
+          PrettyPrinter(System.out)(s.value)
+          println
+          println
+          val (c, v) = s.value.eval.run(Fixture.configuredContext)
+          println("Expanded to:")
+          PrettyPrinter(System.out)(ast.SBFile(v))
+          println
+          println
+          println("With logs:")
+          println(c.logms.map(_.pretty).mkString("\n"))
+          val doc = Exporter[datatree.ast.AstDatatree](c).apply(v)
+          val rdfIo = implicitly[RdfIo[datatree.ast.AstDatatree]]
+          val xmlWriter = new IndentingXMLStreamWriter(
+            XMLOutputFactory.newInstance.createXMLStreamWriter(
+              new FileWriter(s"$file.rdf")))
+          rdfIo.write(xmlWriter, doc)
+        case e : Failure =>
+          System.err.println(s"Problem parsing $file")
+      }
+    } else {
+      System.err.println(s"File $file doesn't exist")
     }
   }
 }
